+++ src/app/shops/[name]/offers/[id]/page.tsx
@@ -1,9 +1,9 @@
-import styles from '@/styles/app/organizations/[name]/offers/[id]/page.module.scss';
+import styles from '@/styles/app/shops/[name]/offers/[id]/page.module.scss';
 import { TopBar } from '@/components/common/top-bar';
 import { PageBox } from '@/app/page-box';
 import { Footer } from '@/components/common/footer';
-import { OfferCodeContainer } from '@/app/organizations/[name]/offers/[id]/offer-id-container';
-import { OfferCodeForm } from '@/app/organizations/[name]/offers/[id]/offer-code-form';
+import { OfferCodeContainer } from '@/app/shops/[name]/offers/[id]/offer-id-container';
+import { OfferCodeForm } from '@/app/shops/[name]/offers/[id]/offer-code-form';
 import { ProfileProvider } from '@/contexts/profile-context';
 
 interface OfferIdPageProps {